not trading any trade which is NEW

+++ trade/serialize/trade_deserializer.py
@@ -1,3 +1,4 @@
+from core.number.BigFloat import BigFloat
 from core.trade.InstrumentTrade import InstrumentTrade, Status
 from utility.json_utility import as_data
 from utility.string_utility import is_empty
@@ -6,7 +7,7 @@ from utility.string_utility import is_empty
 def deserialize_trade(trade) -> InstrumentTrade:
     instrument_from = as_data(trade, 'instrument_from')
     instrument_to = as_data(trade, 'instrument_to')
-    quantity = as_data(trade, 'quantity')
+    quantity = BigFloat(as_data(trade, 'quantity'))
     status = as_data(trade, 'status')
     description = as_data(trade, 'description')
     deserialized_trade = InstrumentTrade(instrument_from, instrument_to, quantity)
